feat(lab/10): run the branch generator in a worker
