+++ tsconfig.json
@@ -9,7 +9,8 @@
       "es2015",
       "es2016",
       "es2017",
-      "scripthost"
+      "scripthost",
+      "webworker"
     ],
     "module": "es2015",
     "moduleResolution": "node",
